+++ src/connectedvmware/azext_connectedvmware/_help.py
@@ -408,13 +408,13 @@ helps[
         text: |-
                az connectedvmware vm create \
 --machine-id /subscriptions/01234567-0123-0123-0123-0123456789ab/resourceGroups/contoso-rg/providers/Microsoft.HybridCompute/machines/contoso-vm \
---vcenter /subscriptions/fedcba98-7654-3210-0123-456789abcdef/resourceGroups/contoso-rg/providers/Microsoft.HybridCompute/vcenters/contoso-vcenter
+--vcenter /subscriptions/fedcba98-7654-3210-0123-456789abcdef/resourceGroups/contoso-rg/providers/Microsoft.ConnectedVMwareVsphere/vcenters/contoso-vcenter
 
       - name: Link an HCRP Machine to a vCenter in another subscription using the machine name.
         text: |-
                 az connectedvmware vm create \
 --resource-group contoso-rg --name hcrp-contoso-machine \
---vcenter /subscriptions/fedcba98-7654-3210-0123-456789abcdef/resourceGroups/contoso-rg/providers/Microsoft.HybridCompute/vcenters/contoso-vcenter
+--vcenter /subscriptions/fedcba98-7654-3210-0123-456789abcdef/resourceGroups/contoso-rg/providers/Microsoft.ConnectedVMwareVsphere/vcenters/contoso-vcenter
 
       - name: Link an HCRP Machine to a vCenter in the same subscription and resource group.
         text: |-
@@ -432,19 +432,19 @@ helps[
         text: |-
                 az connectedvmware vm create-from-machines \
 --resource-group contoso-rg --name contoso-vm \
---vcenter-id /subscriptions/fedcba98-7654-3210-0123-456789abcdef/resourceGroups/contoso-rg-2/providers/Microsoft.HybridCompute/vcenters/contoso-vcenter
+--vcenter-id /subscriptions/fedcba98-7654-3210-0123-456789abcdef/resourceGroups/contoso-rg-2/providers/Microsoft.ConnectedVMwareVsphere/vcenters/contoso-vcenter
 
       - name: Creates VMware resources from all Arc for Servers machines in the specified resource group belonging to that vCenter
         text: |-
                 az connectedvmware vm create-from-machines \
 --resource-group contoso-rg \
---vcenter-id /subscriptions/fedcba98-7654-3210-0123-456789abcdef/resourceGroups/contoso-rg-2/providers/Microsoft.HybridCompute/vcenters/contoso-vcenter
+--vcenter-id /subscriptions/fedcba98-7654-3210-0123-456789abcdef/resourceGroups/contoso-rg-2/providers/Microsoft.ConnectedVMwareVsphere/vcenters/contoso-vcenter
 
       - name: Create VMware resources from all Arc for Servers machines in the specified subscription belonging to that vCenter
         text: |-
                 az connectedvmware vm create-from-machines \
 --subscription contoso-sub \
---vcenter-id /subscriptions/fedcba98-7654-3210-0123-456789abcdef/resourceGroups/contoso-rg-2/providers/Microsoft.HybridCompute/vcenters/contoso-vcenter
+--vcenter-id /subscriptions/fedcba98-7654-3210-0123-456789abcdef/resourceGroups/contoso-rg-2/providers/Microsoft.ConnectedVMwareVsphere/vcenters/contoso-vcenter
 """
 
 helps[
@@ -670,8 +670,14 @@ helps[
       - name: Enable guest agent on the vm
         text: |-
                az connectedvmware vm guest-agent enable --username "vm user name" --password "vm password" \
-               --resource-group "resource group name" --subscription "Name or ID of subscription" \
-               --vm-name "vm name"
+--resource-group "resource group name" --subscription "Name or ID of subscription" --vm-name "vm name"
+
+      - name: Enable guest agent on the vm using private link
+        text: |-
+               az connectedvmware vm guest-agent enable --username "vm user name" --password "vm password" \
+--resource-group "resource group name" --subscription "Name or ID of subscription" --vm-name "vm name" \
+--private-link-scope "/subscriptions/fedcba98-7654-3210-0123-456789abcdef/resourceGroups/contoso-rg/providers/Microsoft.HybirdCompute/privateLinkScopes/contoso-private-link-scope"
+      
 """
 
 helps[